Apply suggested changes. Fix cspell.
